fix: register contributions from services files
so importing `vscode/monaco` isn't mandatory
