+++ src/services.ts
@@ -1,4 +1,5 @@
 import './missing-services'
+import './contributions'
 import Severity from 'vs/base/common/severity'
 import { IConfigurationChangeEvent } from 'vs/platform/configuration/common/configuration'
 import { ITextModelContentProvider } from 'vs/editor/common/services/resolverService'
